preserve journal when going from persistent to volatile and back

+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/scripts/journal-set-volatile.sh
@@ -1,5 +1,19 @@
-systemctl stop systemd-journald
-rm -rf /var/log/journal
+#!/bin/bash
+
+if grep -qs -e '^Storage=volatile' /etc/systemd/journald.conf; then
+    echo set-volatile.sh: already volatile, no actions performed
+    exit 0
+fi
+
+# move over the existing log, if it doesn't work so be it
+mkdir -p /run/log/journal
+if [[ $(du -s "/var/log/journal/$(cat /etc/machine-id)" | cut -f1) < 20000 ]]; then
+    cp -f -a "/var/log/journal/$(cat /etc/machine-id)" /run/log/journal/
+fi
+
 sed -i -e 's/.*Storage=.*/Storage=volatile/' "/etc/systemd/journald.conf"
 sed -i -e 's/.*RuntimeMaxUse=.*/RuntimeMaxUse=10M/' "/etc/systemd/journald.conf"
+
 systemctl restart systemd-journald
+
+rm -rf /var/log/journal
